#231 Create Applicant API change  for sunlight

+++ DataReef.TM.Models/DTOs/Solar/Finance/Applicants.cs
@@ -57,7 +57,7 @@ namespace DataReef.TM.Models.DTOs.Solar.Finance
         public bool isSyncing { get; set; }
         public string quoteNumber { get; set; }
         public DateTime createdDateTime { get; set; }
-        public int loanAmount { get; set; }
+        public decimal loanAmount { get; set; }
         public double finalMonthlyPayment { get; set; }
         public double finalEscalatedMonthlyPayment { get; set; }
         public double monthlyPayment { get; set; }
